+++ packages/react-wildcat-handoff/test/serverHandoffSpec.js
@@ -353,6 +353,33 @@ describe("react-wildcat-handoff/server", () => {
             });
         });
 
+        context("handles non-aliased subdomains", () => {
+            ["async", "sync"].forEach((timing) => {
+                it(timing, (done) => {
+                    const serverHandoff = server(stubs.domainsWithoutAliasedSubdomains[timing]);
+
+                    expect(serverHandoff)
+                        .to.be.a("function")
+                        .that.has.property("name")
+                        .that.equals("serverHandoff");
+
+                    const result = serverHandoff(stubs.requests.noSubdomain, stubs.cookieParser, stubs.wildcatConfig)
+                        .then(response => {
+                            expect(response)
+                                .to.be.an("object")
+                                .that.has.property("html")
+                                .that.is.a("string");
+
+                            done();
+                        })
+                        .catch(error => done(error));
+
+                    expect(result)
+                        .to.be.an.instanceof(Promise);
+                });
+            });
+        });
+
         context("matches domains", () => {
             ["async", "sync"].forEach((timing) => {
                 it(timing, (done) => {
@@ -364,6 +391,33 @@ describe("react-wildcat-handoff/server", () => {
                         .that.equals("serverHandoff");
 
                     const result = serverHandoff(stubs.requests.basic, stubs.cookieParser, stubs.wildcatConfig)
+                        .then(response => {
+                            expect(response)
+                                .to.be.an("object")
+                                .that.has.property("html")
+                                .that.is.a("string");
+
+                            done();
+                        })
+                        .catch(error => done(error));
+
+                    expect(result)
+                        .to.be.an.instanceof(Promise);
+                });
+            });
+        });
+
+        context("matches domain aliases", () => {
+            ["async", "sync"].forEach((timing) => {
+                it(timing, (done) => {
+                    const serverHandoff = server(stubs.domains[timing]);
+
+                    expect(serverHandoff)
+                        .to.be.a("function")
+                        .that.has.property("name")
+                        .that.equals("serverHandoff");
+
+                    const result = serverHandoff(stubs.requests.ip, stubs.cookieParser, stubs.wildcatConfig)
                         .then(response => {
                             expect(response)
                                 .to.be.an("object")