+++ CHANGELOG.md
@@ -1,3 +1,33 @@
+<a name="4.4.0"></a>
+# [4.4.0](https://github.com/nfl/react-wildcat/compare/4.3.1...v4.4.0) (2016-07-22)
+
+
+### Features
+
+* **react-wildcat-handoff:** Ability to alias by subdomain ([#146](https://github.com/nfl/react-wildcat/issues/146)) ([6221547](https://github.com/nfl/react-wildcat/commit/6221547))
+
+
+
+<a name="4.3.1"></a>
+## [4.3.1](https://github.com/nfl/react-wildcat/compare/4.3.0...v4.3.1) (2016-07-18)
+
+
+### Bug Fixes
+
+* **react-wildcat-handoff:** Replace arrow functions with named callbacks. ([9dd5f5f](https://github.com/nfl/react-wildcat/commit/9dd5f5f))
+
+
+
+<a name="4.3.0"></a>
+# [4.3.0](https://github.com/nfl/react-wildcat/compare/4.2.0...v4.3.0) (2016-07-13)
+
+
+### Features
+
+* **react-wildcat-handoff:** Add support for domain aliases ([#144](https://github.com/nfl/react-wildcat/issues/144)) ([249afcd](https://github.com/nfl/react-wildcat/commit/249afcd))
+
+
+
 <a name="4.2.0"></a>
 # [4.2.0](https://github.com/nfl/react-wildcat/compare/4.1.0...v4.2.0) (2016-06-27)